+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 6.7
 --------------
 - Fix PHP warning array_values() expects parameter 1 to be array, null given in Charts.fnc.php
+- Group Periods & remove `_getPeriod()` in MasterScheduleReport.php
 
 Changes in 6.6.1
 ----------------